Add a documentation reporter
Similar to rspec, list a hierarchical overview of all testing contexts.

This commit also changes where the default config is merged in, so that
kaocha.test/run can be called without having to specify all defaults.

Add better error reporting when the reporter value is bogus.

Make ns-patterns work with strings, this way we don't have to deal with the
weirdness that ensues because regexes don't have value semantics.