Print test results and failure info when process is interrupted
When interrupting the process with Ctrl-C before the test run is complete,
finish with printing all results collected so far before exiting.

Also fix --fail-fast, so that it bails immediately when an assertion fails,
rather than at the end of the var.

Correctly report number of test vars + number of assertions, rather than taking
the number of assertions for the number of test vars.