Implement random order and seeding
Shuffle namespaces and vars up based on a specific random seed. The seed can be
reused to recreate the same test order.

This helps detect subtle ordering issues where test A only passes when test B
runs first. That this is no idle claim is demonstrated by the fact that this
commit already contains a fix to such an issue, one of the test loading tests
was assuming a test namespace was already loaded before the test had begun.