Implement file system watching
Use hawk to watch source-paths and test-paths for changes, reload the files, and
re-run tests. This does pull in a few extra deps (hawk obviously, and also
core.async, cause I couldn't figure out a better way to respond to multiple file
change events while only having a single test run happening at once).

This also introduces the source-paths config key, which is neat to have anyway
because it also allows us e.g. to integrate with cloverage. I'm still going on
the assumption though that you're configuring your source paths in deps.edn, so
we're not automatically adding these to the classpath, in contrast to the
test-paths.

Significantly expanded the README.md to point out the more salient points
regarding configuring and using Kaocha.