Namespace keys in configuration
This is a non-backwards compatible change, all keys in configuration maps are
namespaced with :kaocha. This is an intermediate step, the next step will be to
distinguish suite-level config by namespacing it with :kaocha.suite.

This enables richer validation with clojure.spec, which in turn with expound can
give us better reporting of faulty configurations.

It's also a prepratory step on the way to the recursive (i.e. generic,
pluggable) loader/runner.