First bit of the reimplementation as kaocha.api
This is a second iteration on the recursive test runner I spiked earlier, this
time driven by specs and tests, and built bottom-up.

It introduces the three core data structures: config, test plan, result, and the
two multimethods that are at the hard of the transformation from each type to
the next, load-testable and run-testable.

CONFIG ---[load-testables]--> TEST PLAN --[run-testables]--> RESULT

Testables are nested and extensible, out of the box Kaocha will come with a
:kaocha.type/suite, which contains :kaocha.type/ns, which contains
:kaocha.type/var.