Load testable implementations lazily based on the type name.
This is convenient both for us (we don't need to require namespaces just for the
multimethods), but even more so for third party contributors. This change makes
it possible to implement a new type simply by adding the right location to the
classpath. Then if a users refers to it in their config, it will automatically
get loaded.