Implement testable/run for suites, plus sketch out api/run
Suite is running, nice! Now we "just" need the command line api wrapper around
it, and an ergonomic command line tool that uses it.

Instead of "middleware" as such decided to go with a "plugin" pattern, that
maybe leans slightly on pedestal's interceptor chain concept, in that plugins
that come later can modify the plugin stack.

Randomize as a plugin seems promising, and shows nicely the benefit of a
data-first approach.

Started putting some stuff in kaocha.api just to think through how it'll fit
together, but that's more a sketch than anything, and doesn't in fact run at the moment.