Add skip/focus flags and config keys
Make it possible to skip specicifc tests, or to run only specific tests, either
based on their id, or on metadata on the test (var, ns).

This can be specified on the command line, or in the config, on the top level or
on individual tests/suites.

Filters are symbols, although specifying them as keywords also works. They need
to either fully match the id of the test, or fully match the namespace part of
the id. Meta filters apply if the given metadata key is truthy.