Make filtering work at the top level
The filter code treated the whole test plan as a testable, but ::testable/skip
is not honored at the top level, a test-plan resembles a test plan but isn't one
and should not be treated as one. Instead filter the top-level :kaocha/tests one
by one.

The immediate impact of this is that if a --focus rule doesn't match any tests,
it will now run 0 tests, instead of all tests.