Provide :kaocha.var/wrap, and correctly handle once-fixture-fn
Provide an official way to wrap test functions, which also makes it possible to
decide if you want to wrap on the inside (inside any each-fixtures), or all
around any other wrappings.

Handle the case of once-fixtures that don't return the result of
calling (tests). This is a fully ok thing to do, but it means some extra book
keeping to make sure we don't lose the ns results.