When fail-fast is true, bail out immediately on load error
Load errors are marked as a failure on the respective namespace, and reported as
a test failure when that namespace runs. This means all load errors are signaled
as warnings before the tests start to run.

Instead, when fail-fast is true, we bail out immediately upon the first load
error.