Fix classloader hack
In order to be able to reliably add directories to the classpath and have
clojure notice them we add them to the Compiler/LOADER, binding Compiler/LOADER
first if it isn't bound already.

The previous version would use push-thread-bindings for this, but since this
push didn't have a corresponding pop it would mess up the stack, so that e.g.
after a while *out* no longer pointed to where you would it expect it to.

Instead simply bind the root if it's not bound already. Now this is also
automatically done, instead of on request.