Improve matcher-combinators support
In the end the only way forward was to implement our own :mismatch handlers.
This allows us to defer printing failures until the summary, and it gives us an
opportunity the print the captured output.

Started using keyword inheritance to clean up some of the reporter stuff and
make it more extensible, with this all the MC stuff is in a single
namespace, *except* for the t/do-report monkey patch, and the history tracker
summary.