When loading a clojure.test ns, don't reload
This means the load steps uses any existing process/REPL state, rather than
force-refreshing namespaces. This is an improvement because `(require :reload)`
is naive anyway, proper reloading needs to be done with tools.namespace, which
is what already happens in watch mode.

When working on the REPL is it very unintuitive that calling `run-tests` will
force-reload the namespace you are working in. You expect to be able to eval
things without saving the file, and have that state be visible. Hence this
change.