Introduce/support Cucumber test runner
This is a large commit with a bunch of changes that fell out of implementing the
Cucumber test type. This is the first main type after clojure.test, so it's been
interesting to find ways to make the current abstractions more generic and
divorced from clojure.test details like namespaces and vars.

This commit also starts the work of converting the old bash-based "feature"
tests to Cucumber/Gherkin features.

Some of the stuff that's in here

- Explicitly give tests a description key, rather than deriving it from var/ns

testables now have a `:kaocha.testable/desc` which can be displayed to the user
e.g. in documentation output.

- Use orchestra+expound for nicer spec errors everywhere

A no-brainer

- Start porting feature tests to Cucumber

As mentioned, there are two features so far under test/features.

- Allow short names for built-in plugins

We already did this for reporters (e.g. bin/kaocha --reporter documentation),
now you can also do it for plugins (--plugin junit-xml).

- Make sure the color option is honored also for reporter resolution errors

This error message was being printed outside of the `binding` that influences
whether kaocha.output uses color.

- Introduce more generic test event types (begin-test-group / begin-test)

Continuing the trend of splitting keywords into hierarchies and working with
higher level generic terms.

- Allow events to be printed only at the final summary (deferred)

Another thing you can mark through hierarchy/derive.

- Introduce the concept of "pending" tests

We have pending and skip. Skip(ped) tests are completely ignored. Pending tests
are reported as such during every test run.

- Provide some conveniences for implementing test types (kaocha.type)

We'll need more of these, but it's a start.

- Add a (substring? ...) assertion

Provide a more helpful message than (not (.contains s1 s2))