Consider the global hierarchy for hierarchy/isa?
This is so that third-party test reporters can register their event types as
being kaocha/known-key or kaocha/fail-type, without having to depend on Kaocha
itself.

Also adds a helper for checking if a testable is a leaf type. Leaf types can not
have children, e.g. kaocha.type/var is a leaf type. Non-leaf types are
containers for tests, e.g. a test-suite or kaocha.type/ns are non-leaf types.