Consider equality assertions that don't compare anything as failures
An assertion of the form `(is (= arg))` will always evaluate to true. Most
likely it's actually a typo, e.g. I have more than once written something like
this.

(is (= {:foo :bar})
    (funtion-call))

Kaocha will now flag this as a failure, similar to how tests without assertions
are flagged as failures.