Make test failure location directly settable by test type
Where possible or appropriate the test type can now directly bind a dynamic
var (testable/*test-location*), to specify the location where a failure
happened. This will override the stacktrace-based lookup.

This is useful mainly in two scenarios: one is where test specification does not
correspond one-to-one with executable code, and thus with a stracktrace element,
e.g. in case of Cucumber.

The other case is when the test type reports a failure outside of the actual
test run function, e.g. when reporting a "zero-assertions" failure after a test
has run.