Add :kaocha.testable/skip-remaining? as an explicit opt-in fail-fast
When a testable's `run` invocation returns a result testable with this flag set,
then any remaining tests at the current level (e.g. any remaining vars in the
namespace) are skipped.

This is intended for cases where a test fails in such a way that it no longer
makes sense to execute the rest. E.g. when a ClojureScript test times out, then
the ClojureScript REPL will no longer be responsive for now commands, so if the
remaining tests would run they would likely just time out themselves.