Filter out jdk.internal.reflect.* from stackframes
When detecting the file/line of a failure, we first look for an file/line set in
the failure itself (because the test type knows best), then try to infer it from
the stacktrace (because this will be the most accurate, i.e. point at the
assertion itself), and finally fall back to the file/line as found on the
test (var) metadata. This should always work but is less accurate.

To do the stack based detection we take the topmost stackframe that isn't system
or kaocha code, so filtering out Clojure, Java, and Kaocha internals.

Since Java 9 the sun.reflect package has been renamed to jdk.internal.reflect,
so we need to filter those out as well.