First steps for a kaocha.test namespace
The idea is that this will be a drop-in replacement for clojure.test, that fixes
some long time annoyances.

For now it's just a single `deftest` macro, that works like the original, but
when evaluated outside of a Kaocha run it will also kick off Kaocha to run the
defined test, providing a smooth REPL workflow.