Add a notifier plugin, based on @rgm's initial implementation
Add a plugin that shells out to `notify-send` or `terminal-notifier` to pop up a
desktop notification bubble when a test run passes or fails.

The command can be configured as shown in the feature test. If neither
notify-send or terminal-notifier is found than it does nothing.

I've added the Clojure logo as a resource and added some code to extract it from
the jar as necessary, since not all notification servers accept URLs for the
icon.

Some other things that snuck in here:

The `defplugin` macro now generates multiple vars, one for each hook, one with
the mapping of hooks, and one for the actual `(defmethod -register ...)`. This
makes it a lot more straightforward to test individual hooks.

The progress bar reporter has been fixed, it didn't make it to 100% when there
were pending tests, this was in part because pending tests didn't generate a
start/end event, only the pending event, but also because the progress bar
reporter predated much of `kaocha.hierarchy`.