A boatload of improvements to --watch
I've tried to address as many of the concerns and issues with --watch as
possible. In the process I've cleaned up the code quite a bit, but it's also
gotten quite a bit more complex, and it's still basically untested because this
stuff is a pain to test, but I've tried to dogfood it as much as possible.

The big issue is that is fixed is that load errors are handled much better. Now
if a namespace fails to load you'll get to see the relevant exception and trace,
and the whole thing is reported as if it's a failing test run. That way plugins
and reporters like the notifier plugin can also report on it. We no longer try
to do an actual run in this case, all tests are skipped.

Pressing enter will re-run all tests (not just previously failing tests).

You can set patterns for files to ignore in `tests.edn` (:kaocha.watch/ignore
[...]). These follow the glob spec of java's PatchMatcher. The original idea was
to source these from .gitignore, but .gitignore uses a pretty specific format,
so we'd have to parse and convert that, that's something for later.

Finally `tests.edn` will also be watched, and it gets reloaded on every run, so
you can change your config and add plugins on the fly.