Remove binding plugin, report load errors
The binding plugin has been removed in favor of making the functionality
built-in. This way we can make sure bindings are installed during the complete
Kaocha process, including the load step.

Errors during the load step now cause the test run to fail. They are still
reported as if a test run happened, so plugins can pick up on them.