Implement :kaocha.testable/wrap for STC
This is a special key that plugins can add to leaf-type testables, it is a
sequence of wrapping functions that get wrapped around the inner most testing
logic. This is to set up bindings that are active while test logic is running,
but not e.g. when emitting :begin / :end events.

In particular this is necessary to make output capturing work correctly.