Fix issues with --watch
The code that watches your tests.edn for changes would fail when tests.edn
didn't exist, or wasn't referenced in the config (through ::cli-options).

This wasn't diagnosed earlier because the exceptions happened in a future and
thus were being lost.

Handle this both in the watch code, by making it check for the file, and in
`kaocha.repl/config` by making sure it sets `:kaocha/cli-options :config-file`.

Closes #97
Closes #93