Error bad ns forms as errors, instead of ignoring
When a test namespace has an ns form that can't be read by tools.namespace, then
tools.namespace would simply pretend that ns did not exist, and we would happily
carry on from there. This is problematic, people should be notified when we are
unable to load a test file.

We've updated our fork of tools.namespace to return these errors including the
file name context, which we then use to turn them into load errors.

When a load error is encountered the test is not run, but it's reported as an
error through the configured reporter. Any sibling tests are skipped.