filter plugin: honor :kaocha.testable/aliases
Currently --focus/--skip looks at the :kaocha.testable/id, but these are not
always convenient to work with. In clojure.test they correspond with
namespace/var names, but for cljs we prefix them with `cljs:` to make sure they
are unique in case of cljc files. In kaocha-cljs2 we will prefix them with a
runtime-specific generated prefix because you might be running them in multiple
runtimes at the same time, in which case each of them gets its own id.

With aliases we can still put the unqualified namespace/var name as an alias, so
people can use --focus/--skip the way they would expect.