Run post-summary hook from kaocha.repl, change test-seq
The post-summary hook is where extra end-of-test-run reporting is done, like
printing profiling information. This hooks does not run when invoking kaocha.api
directly, because we don't want it to run when using kaocha as a library. We do
want it to run when invoked by an end user, so through kaocha.runner, or through
kaocha.repl, so it makes sense to invoke it in both of those places.

Also includes a change to test-seq, a function which flattens a hierarchy of
testables into a single seq. When called on something that isn't a testable like
a config map or test-plan, it would include that outer container in the seq as
well. This is generally not desirable. This is an internal change, but could
potentially impact third party plugins that use test-seq.