Orchestra plugin: render explain-str as message
This is to accomodate a breaking change in Orchestra. Before it would add the
explain string to the exception, but this is no longer case. Instead you are
expected to take the exception's ex-data and render the explain data yourself.

Also be a bit more selective when calling `orchestra/instrument` for ourselves:
only instrument kaocha/lambdaisland code. If people want to instrument in
general they should use the orchestra plugin.