Pre-filter test suites to possible skip loading of suites
Filtering via focus/skip(-meta) directives is done by walking the test plan,
this means that all test suites are loaded, and only then is filtering applied.

The result is that all test suites are always loaded, even when some will not be
run.

This change adds an extra filtering pass before the load stage to see if we
already have enough information to mark test suites as to be skipped, so that
they don't get loaded. This only works if focus/skip is applied on the test
suite itself (e.g. :unit, :integration), rather than on a specific test group or
test within that suite.