Check for kaoch/cli-args in the spec-test plugin
This plugin generates and injects a synthetic test suite, it does so in a config
hook, but the config hook is called after command line arguments are applied.
This means that it bypasses test suite selection, so it is impossible to run
only this test suite from the command line, and it causes this generated suite
to run even if a different suite was requested.

Reshuffling the order of config/hook handling could cause other changes in
behavior, so instead opting to have the plugin handle this case. This will be
generally recommeded: if plugins generate test suites, they should check
kaocha/cli-args themselves.