Skip running of empty test groups
When running a group-level test (e.g. namespace) check first if there are any
actual leaf tests in there that will run (that are not already marked as
skipped). Otherwise skip the group test.

This is mainly for the spec-test plugin, which will create test groups for every
NS, even if none of the functions in the NS have fdefs.