Clean up load-error handling and fix handling in kaocha.watch
This makes kaocha.watch use the mechanisms already available in
`kaocha.testable` to deal with and report load errors. It also does a general
pass of cleaning things up, because we had created a bit of a mess.

`handle-load-error` is gone, the testable implementations should no longer need
this, just let kaocha.testable handle this stuff for you.

It also seems at some point we went from `:kaocha.test-plan/load-error` to
`:kaocha.testable/load-error`, and ended up with a bit of a hodge-podge. It's
surprising things still worked as well as they did. I changed the remaining
test-plan/load-error references to testable/load-error, since that now
dominates, and is IMO also the more intuitive, since we apply these on
individual testables (typically the suite).