Apply bindings higher in the stack, check if history var is bound
Add an extra (with-bindings (config/bindig-map config) ...) at the top level in
kaocha.runner, we add the bindings later in `api/run` as well, and still need to
do that too, in case a plugin modifies the binding map. This new binding
wrapping is useful to influence `:main` and `:post-summary` plugin invocations.

Make the defaults `history/track` implementation more defensive, so it doesn't
throw when called outside the scope where `history/*history*` is bound. This
fixes an unfortunate interaction with spec and test.check.