Don't skip tests with load errors
Currently this leads to an unfortunate interaction where if you `--focus
foo/bar`, but `foo.clj` fails to load, then there is no `#'foo/bar`, and the `foo`
ns gets skipped, even though the `foo` ns actually has a load error which we
need to show the user.

So: never mark testables with load errors as to be skipped, we're better off
being cautious here and always surfacing load-errors.