Split error code 249 into two.
I separated failures to load namespaces from failures to load the
plugin. (The two instances of 248 are separate paths but are only separate so
we can use the correct grammar when there's one failed namespace versus
multiple.)

When registering a plugin fails due to being unable to load a namespace,
the error code is now 249. When registering a plugin fails for other
reasons, the error code is now 248. When resolving a reporter var fails,
the error code is still 254. Users can test for error codes between 240
and 249 to check for plugin errors regardless of cause.