Fix bug in making config keys replace by default.
When combining two configurations, certain keys are meant to replace the
keys in the previous map rather than being merged, unless the user
specifies otherwise. However, in reality, these keys would always be replaced,
regardless of metadata. This bug was probably missed because of how
seldom you want to override the default behavior and because I think
people forget meta-merge is an option.