Test against parsed configuration, not string.
Avoid issues with nondeterminism in the order of configuration map
printing by parsing the configuration as EDN, and then checking for the
contents we expect.

I tried eliminating various parts through trial and error to get the
results to appear as they did when the test was written, but no luck.
It may just be the latest Clojure tooling plus JVM versions after 11.
However, pursuing that rabbit hole further didn't seem worth it for a
guarantee we don't offer—we don't promise that the configuration will
print in a particular order.
(Those changes don't appear in version history because I squashed them.)