Update Beholder to improve support for certain editors.
Beholder released a fix for certain editors that create a new file and
then delete instead of modify. I believe Vim and Neovim do this when
noswapfile and either backupcopy=no or backupcopy=auto are set, but
other editors probably do it too.

Sometimes these creation events would be dropped because the file exists
already. I haven't noticed this problem when using --watch with Neovim,
but it may depend on timing and thus may be highly platform-, version-
or setup-specific. Also, I consider this a safe change to makee without
reproducing the initial issue.