Make integration helpers less likely to hang.
If there's nothing left to slurp, expect-lines and next-line-matches
would previously just hang, which slows you down and leaves you with a
mystery. What were the final things emitted by the integration test
before it all went wrong, you might wonder. Alas, they are lost to time.

By checking the ready method, we can only try to slurp from *in* if
there's something to slurp. If there isn't, we print what we tried to
match against.

I don't know if there are some situations where the process would still
be unspooling information, and blocking on slurp for a brief moment
allows for capturing that information. I think missing out on that is
acceptable since you are already in a failure state and hanging is muc
worse. Adding a short delay could mitigate this, however.