Rudimentary fix to get integration test to pass again.
The primary change is to relax checks by using regexes instead. This is
perhaps not the most elegant solution, as the regexes are pretty
ungainly and not necessarily trivial to interpret or change. On the
other hand, it makes the test robust to some non-determinism and allows
for similar changes in the future, without completely redesigning
expect-lines or next-line-matches.