Switch from embeds to equals to improve messages
Artificially dividing it up produced weird errors when there was a
mismach. Using equals results in more coherent diffs and only required
adding a few lines.

I also didn't have that much confidence that a different change in test
output wouidn't trigger a long delay.

This does make the test slightly more fragile, as any thing added to the
beginning of the output or the end of the output will cause this to
fail.

Also, if there's more lines in the output than in the test, you don't
get a great diff either.